[rpc] move getinfo implementation to peer_control.c
